+++ src/main/java/svenhjol/charm/module/wandering_trader_maps/WanderingTraderMaps.java
@@ -1,8 +1,8 @@
 package svenhjol.charm.module.wandering_trader_maps;
 
 import net.minecraft.core.BlockPos;
-import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.npc.VillagerTrades;
@@ -13,8 +13,8 @@ import net.minecraft.world.level.saveddata.maps.MapDecoration;
 import svenhjol.charm.Charm;
 import svenhjol.charm.annotation.CommonModule;
 import svenhjol.charm.annotation.Config;
+import svenhjol.charm.helper.LogHelper;
 import svenhjol.charm.helper.MapHelper;
-import svenhjol.charm.helper.StringHelper;
 import svenhjol.charm.helper.VillagerHelper;
 import svenhjol.charm.helper.WorldHelper;
 import svenhjol.charm.loader.CharmModule;
@@ -25,15 +25,15 @@ import java.util.Random;
 
 @CommonModule(mod = Charm.MOD_ID, description = "Wandering traders have a chance to sell maps to distant structures.")
 public class WanderingTraderMaps extends CharmModule {
-    @Config(name = "Structures", description = "List of structure IDs and structure tags.")
+    @Config(name = "Maps to structures", description = "List of structure tags.")
     public static List<String> structures = Arrays.asList(
-        "#minecraft:village", "#minecraft:mineshaft", "minecraft:swamp_hut", "minecraft:igloo"
+        "minecraft:village", "minecraft:mineshaft", "charm:swamp_hut", "charm:igloo"
     );
 
-    @Config(name = "Rare structures", description = "List of rarer structure IDs and structure tags.\n" +
+    @Config(name = "Maps to rare structures", description = "List of rarer structure tags.\n" +
         "These are more expensive and less likely to be sold.")
     public static List<String> rareStructures = Arrays.asList(
-        "minecraft:desert_pyramid", "minecraft:jungle_temple", "#minecraft:ocean_ruin"
+        "charm:desert_pyramid", "charm:jungle_temple", "minecraft:ocean_ruin"
     );
 
     @Override
@@ -47,11 +47,25 @@ public class WanderingTraderMaps extends CharmModule {
     }
 
     public static ItemStack makeTraderMap(String id, ServerLevel level, BlockPos pos, int distance, int color) {
-        var nearest = WorldHelper.findNearestMapFeature(id, level, pos, distance, true);
+        if (id.startsWith("#")) {
+            // Differentiating tags and raw strings is now deprecated - we only use tags now.
+            id = id.substring(1);
+        }
+        var res = new ResourceLocation(id);
+
+        LogHelper.debug(WanderingTraderMaps.class, "WT wants to sell: " + res);
+        var nearest = WorldHelper.findNearestStructure(res, level, pos, distance, true);
         if (nearest == null) return new ItemStack(Items.MAP);
 
-        var name = StringHelper.snakeToPretty(id.substring(id.indexOf(":") + 1), true);
-        var mapName = new TranslatableComponent("filled_map.charm.trader_map", new TextComponent(name));
+        var namespace = res.getNamespace();
+        var path = res.getPath();
+        var name = new TranslatableComponent("structure." + namespace + "." + path);
+
+        if (name.getString().contains(".")) {
+            name = new TranslatableComponent("filled_map.charm.structure");
+        }
+
+        var mapName = new TranslatableComponent("filled_map.charm.trader_map", name);
         return MapHelper.create(level, nearest, mapName, MapDecoration.Type.TARGET_X, color);
     }