+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed disabling amethyst noteblock causing IllegalArgumentException crash (#745)
 * Changed Automatic Recipe Unlock to be disabled by default.
+* Wandering Trader maps now only use structure tags and map names are now localised.
 
 ## 4.1.2